+++ xmtp_mls/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod address;
 pub mod hash;
 pub mod id;
 #[cfg(test)]
